生成编号

+++ core6_MVC/coreWeb_MVC/Controllers/UsersController.cs
@@ -1,5 +1,4 @@
-﻿#nullable disable
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -22,19 +21,21 @@ namespace coreWeb_MVC.Controllers
         // GET: Users
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Users.ToListAsync());
+              return _context.Users != null ? 
+                          View(await _context.Users.ToListAsync()) :
+                          Problem("Entity set 'TestDBContext.Users'  is null.");
         }
 
         // GET: Users/Details/5
-        public async Task<IActionResult> Details(int? id)
+        public async Task<IActionResult> Details(string id)
         {
-            if (id == null)
+            if (id == null || _context.Users == null)
             {
                 return NotFound();
             }
 
             var user = await _context.Users
-                .FirstOrDefaultAsync(m => m.ID == id);
+                .FirstOrDefaultAsync(m => m.UserID == id);
             if (user == null)
             {
                 return NotFound();
@@ -54,7 +55,7 @@ namespace coreWeb_MVC.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,UserID,Password,UserName,Sex,Age,Introduce,Address")] User user)
+        public async Task<IActionResult> Create([Bind("UserID,Account,Password,UserName,UserEmail,Sex,Age,Introduce,Address,Birthday,CreatedAt,UserType,UserState,AddDate")] User user)
         {
             if (ModelState.IsValid)
             {
@@ -66,9 +67,9 @@ namespace coreWeb_MVC.Controllers
         }
 
         // GET: Users/Edit/5
-        public async Task<IActionResult> Edit(int? id)
+        public async Task<IActionResult> Edit(string id)
         {
-            if (id == null)
+            if (id == null || _context.Users == null)
             {
                 return NotFound();
             }
@@ -86,9 +87,9 @@ namespace coreWeb_MVC.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,UserID,Password,UserName,Sex,Age,Introduce,Address")] User user)
+        public async Task<IActionResult> Edit(string id, [Bind("UserID,Account,Password,UserName,UserEmail,Sex,Age,Introduce,Address,Birthday,CreatedAt,UserType,UserState,AddDate")] User user)
         {
-            if (id != user.ID)
+            if (id != user.UserID)
             {
                 return NotFound();
             }
@@ -102,7 +103,7 @@ namespace coreWeb_MVC.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!UserExists(user.ID))
+                    if (!UserExists(user.UserID))
                     {
                         return NotFound();
                     }
@@ -117,15 +118,15 @@ namespace coreWeb_MVC.Controllers
         }
 
         // GET: Users/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(string id)
         {
-            if (id == null)
+            if (id == null || _context.Users == null)
             {
                 return NotFound();
             }
 
             var user = await _context.Users
-                .FirstOrDefaultAsync(m => m.ID == id);
+                .FirstOrDefaultAsync(m => m.UserID == id);
             if (user == null)
             {
                 return NotFound();
@@ -137,17 +138,25 @@ namespace coreWeb_MVC.Controllers
         // POST: Users/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(string id)
         {
+            if (_context.Users == null)
+            {
+                return Problem("Entity set 'TestDBContext.Users'  is null.");
+            }
             var user = await _context.Users.FindAsync(id);
-            _context.Users.Remove(user);
+            if (user != null)
+            {
+                _context.Users.Remove(user);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
-        private bool UserExists(int id)
+        private bool UserExists(string id)
         {
-            return _context.Users.Any(e => e.ID == id);
+          return (_context.Users?.Any(e => e.UserID == id)).GetValueOrDefault();
         }
     }
 }
